Moved people into its own file
