+++ static/js/people/people_test.js
@@ -1,0 +1,2 @@
+require('../../../test/c9test');
+require('../../../test/test.people');
